requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sentence-transformers==2.6.1
faiss-cpu==1.8.0
psutil==5.9.8
pydantic==2.6.4
python-dotenv==1.0.1
google-generativeai==0.5.0  # Gemini API client
redis